+++ src/NUnitFramework/framework/Constraints/DateTimes.cs
@@ -1,5 +1,5 @@
 // ***********************************************************************
-// Copyright (c) 2014 Charlie Poole, Rob Prouse
+// Copyright (c) 2021 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
 // a copy of this software and associated documentation files (the
@@ -8,10 +8,10 @@
 // distribute, sublicense, and/or sell copies of the Software, and to
 // permit persons to whom the Software is furnished to do so, subject to
 // the following conditions:
-//
+// 
 // The above copyright notice and this permission notice shall be
 // included in all copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
 // EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
 // MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -21,10 +21,31 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ***********************************************************************
 
-using System.Reflection;
+using System;
 
-//
-// Current version for the NUnit Framework
-//
-[assembly: AssemblyVersion("3.13.1.0")]
-[assembly: AssemblyFileVersion("3.13.1.0")]
+namespace NUnit.Framework.Constraints
+{
+    /// <summary>
+    /// The DateTimes class contains common operations on Date and Time values.
+    /// </summary>
+    public static class DateTimes
+    {
+        #region DateTimes Difference
+
+        internal static object Difference(object x, object y)
+        {
+            if (x is DateTime && y is DateTime)
+                return ((DateTime)x - (DateTime)y).Duration();
+
+            if (x is TimeSpan && y is TimeSpan)
+                return ((TimeSpan)x - (TimeSpan)y).Duration();
+
+            if (x is DateTimeOffset && y is DateTimeOffset)
+                return ((DateTimeOffset)x - (DateTimeOffset)y).Duration();
+
+            throw new ArgumentException("Both arguments must be DateTime, DateTimeOffset or TimeSpan");
+        }
+
+        #endregion
+    }
+}