+++ src/NUnitFramework/mock-assembly/MockAssembly.cs
@@ -22,6 +22,7 @@
 // ***********************************************************************
 
 using System;
+using System.Collections.Generic;
 using System.Reflection;
 using NUnit.Framework;
 using NUnit.Framework.Internal;
@@ -113,12 +114,12 @@ namespace NUnit.Tests
         [Category("FixtureCategory")]
         public class MockTestFixture
         {
-            public const int Tests = 10;
-            public const int Suites = 1;
+            public const int Tests = 13;
+            public const int Suites = 2;
 
             public const int Passed = 2;
 
-            public const int Skipped_Ignored = 1;
+            public const int Skipped_Ignored = 4;
             public const int Skipped_Explicit = 1;
             public const int Skipped = Skipped_Ignored + Skipped_Explicit;
 
@@ -154,6 +155,19 @@ namespace NUnit.Tests
 
             [Test, Ignore("Ignore Message")]
             public void IgnoreTest() { }
+
+            [TestCaseSource(nameof(SkippedTestCaseData))]
+            public void SkippedTest(int _)
+            {
+                Assert.Pass();
+            }
+
+            public static IEnumerable<TestCaseData> SkippedTestCaseData => new[]
+            {
+                new TestCaseData(1).Ignore("Ignore testcase"),
+                new TestCaseData(2).Ignore("Ignore testcase"),
+                new TestCaseData(3).Ignore("Ignore testcase")
+            };
 
             [Test, Explicit]
             public void ExplicitTest() { }