+++ src/NUnitFramework/tests/Attributes/LifeCycleAttributeTests.cs
@@ -104,6 +104,126 @@ namespace NUnit.Framework.Attributes
         }
 
         [Test]
+        public void InstancePerTestCaseShouldApplyToTestFixtureSourceTests()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(LifeCycleWithTestFixtureSourceFixture));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void InstancePerTestCaseShouldApplyToTestCaseTests()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(FixtureWithTestCases));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void InstancePerTestCaseShouldApplyToTestCaseSourceTests()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(FixtureWithTestCaseSource));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void InstancePerTestCaseShouldApplyToTestsWithValuesParameters()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(FixtureWithValuesAttributeTest));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void InstancePerTestCaseShouldApplyToTheories()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(FixtureWithTheoryTest));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+#if NETFRAMEWORK
+        [Test]
+        public void AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelFixtureLifeCycleTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void GivenFixtureWithTestFixtureSource_AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelLifeCycleTestFixtureSourceTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void GivenFixtureWithTestCases_AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelLifeCycleFixtureWithTestCasesTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void GivenFixtureWithTestCaseSource_AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelLifeCycleFixtureWithTestCaseSourceTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void GivenFixtureWithValuesAttribute_AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelLifeCycleFixtureWithValuesTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void GivenFixtureWithTheory_AssemblyLevelInstancePerTestCaseShouldCreateInstanceForEachTestCase()
+        {
+            var asm = TestAssemblyHelper.GenerateInMemoryAssembly(
+                AssemblyLevelLifeCycleFixtureWithTheoryTest.Code, new[] { typeof(Test).Assembly.Location });
+            var testType = asm.GetType("FixtureUnderTest");
+            var fixture = TestBuilder.MakeFixture(testType);
+
+            ITestResult result = TestBuilder.RunTest(fixture);
+
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+#endif
+
+        [Test]
         public void InstancePerTestCaseWithRepeatShouldWorkAsExpected()
         {
             RepeatingLifeCycleFixtureInstancePerTestCase.RepeatCounter = 0;
@@ -119,6 +239,22 @@ namespace NUnit.Framework.Attributes
         {
             var fixture = TestBuilder.MakeFixture(typeof(ParallelLifeCycleFixtureInstancePerTestCase)); 
             
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void NestedFeatureWithoutLifeCycleShouldInheritLifeCycle()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(LifeCycleWithNestedFixture));
+            ITestResult result = TestBuilder.RunTest(fixture);
+            Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
+        }
+
+        [Test]
+        public void NestedFeatureWithLifeCycleShouldOverrideLifeCycle()
+        {
+            var fixture = TestBuilder.MakeFixture(typeof(LifeCycleWithNestedOverridingFixture));
             ITestResult result = TestBuilder.RunTest(fixture);
             Assert.That(result.ResultState.Status, Is.EqualTo(TestStatus.Passed));
         }